requirements.txt
Werkzeug==0.16.1
markupsafe==2.0.1
flask
folium
pandas
flask_cors
flask_oidc
flask_restx
haversine
PyJWT
jwcrypto
xlsxwriter